move Chunker to its own file; replace BufferedReader with PushbackReader
