+++ Core/src/org/sleuthkit/autopsy/coordinationservice/CoordinationService.java
@@ -1,7 +1,7 @@
 /*
  * Autopsy Forensic Browser
  *
- * Copyright 2015 Basis Technology Corp.
+ * Copyright 2011-2017 Basis Technology Corp.
  * Contact: carrier <at> sleuthkit <dot> org
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,129 +18,89 @@
  */
 package org.sleuthkit.autopsy.coordinationservice;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import org.apache.zookeeper.CreateMode;
-import org.apache.zookeeper.ZooDefs;
 import org.apache.curator.RetryPolicy;
-import org.apache.curator.retry.ExponentialBackoffRetry;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.framework.recipes.locks.InterProcessMutex;
 import org.apache.curator.framework.recipes.locks.InterProcessReadWriteLock;
+import org.apache.curator.retry.ExponentialBackoffRetry;
+import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.KeeperException;
-import org.sleuthkit.autopsy.core.UserPreferences;
-import java.io.IOException;
-import org.apache.zookeeper.WatchedEvent;
-import org.apache.zookeeper.ZooKeeper;
 import org.apache.zookeeper.KeeperException.NoNodeException;
+import org.apache.zookeeper.WatchedEvent;
+import org.apache.zookeeper.ZooDefs;
+import org.apache.zookeeper.ZooKeeper;
+import org.sleuthkit.autopsy.core.UserPreferences;
 
 /**
- * A centralized service for maintaining configuration information and providing
- * distributed synchronization using a shared hierarchical namespace of nodes.
+ * A coordination service for maintaining configuration information and
+ * providing distributed synchronization using a shared hierarchical namespace
+ * of nodes.
+ *
+ * TODO (JIRA 2205): Simple refactoring for general use.
  */
 public final class CoordinationService {
 
-    /**
-     * Category nodes are the immediate children of the root node of a shared
-     * hierarchical namespace managed by the coordination service.
-     */
-    public enum CategoryNode { // RJCTODO: Move this to CoordinationServiceNamespace
-
-        CASES("cases"),
-        MANIFESTS("manifests"),
-        CONFIG("config"),
-        RESOURCE("resource");
-
-        private final String displayName;
-
-        private CategoryNode(String displayName) {
-            this.displayName = displayName;
-        }
-
-        public String getDisplayName() {
-            return displayName;
-        }
-    }
-
-    /**
-     * Exception type thrown by the coordination service.
-     */
-    public final static class CoordinationServiceException extends Exception {
-
-        private static final long serialVersionUID = 1L;
-
-        private CoordinationServiceException(String message) {
-            super(message);
-        }
-
-        private CoordinationServiceException(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
-
-    /**
-     * An opaque encapsulation of a lock for use in distributed synchronization.
-     * Instances are obtained by calling a get lock method and must be passed to
-     * a release lock method.
-     */
-    public static class Lock implements AutoCloseable {
-
-        /**
-         * This implementation uses the Curator read/write lock. see
-         * http://curator.apache.org/curator-recipes/shared-reentrant-read-write-lock.html
-         */
-        private final InterProcessMutex interProcessLock;
-        private final String nodePath;
-
-        private Lock(String nodePath, InterProcessMutex lock) {
-            this.nodePath = nodePath;
-            this.interProcessLock = lock;
-        }
-
-        public String getNodePath() {
-            return nodePath;
-        }
-
-        public void release() throws CoordinationServiceException {
-            try {
-                this.interProcessLock.release();
-            } catch (Exception ex) {
-                throw new CoordinationServiceException(String.format("Failed to release the lock on %s", nodePath), ex);
-            }
-        }
-
-        @Override
-        public void close() throws CoordinationServiceException {
-            release();
-        }
-    }
-
     private static CuratorFramework curator = null;
     private static final Map<String, CoordinationService> rootNodesToServices = new HashMap<>();
-    private final Map<String, String> categoryNodeToPath = new HashMap<>();
     private static final int SESSION_TIMEOUT_MILLISECONDS = 300000;
     private static final int CONNECTION_TIMEOUT_MILLISECONDS = 300000;
     private static final int ZOOKEEPER_SESSION_TIMEOUT_MILLIS = 3000;
     private static final int ZOOKEEPER_CONNECTION_TIMEOUT_MILLIS = 15000;
-    private static final int PORT_OFFSET = 1000;
+    private static final int PORT_OFFSET = 1000; // When run in Solr, ZooKeeper defaults to Solr port + 1000
+    private final Map<String, String> categoryNodeToPath = new HashMap<>();
 
     /**
-     * Gets an instance of the centralized coordination service for a specific
-     * namespace.
+     * Determines if ZooKeeper is accessible with the current settings. Closes
+     * the connection prior to returning.
+     *
+     * @return true if a connection was achieved, false otherwise
+     *
+     * @throws InterruptedException
+     * @throws IOException
+     */
+    private static boolean isZooKeeperAccessible() throws InterruptedException, IOException {
+        boolean result = false;
+        Object workerThreadWaitNotifyLock = new Object();
+        int zooKeeperServerPort = Integer.valueOf(UserPreferences.getIndexingServerPort()) + PORT_OFFSET;
+        String connectString = UserPreferences.getIndexingServerHost() + ":" + zooKeeperServerPort;
+        ZooKeeper zooKeeper = new ZooKeeper(connectString, ZOOKEEPER_SESSION_TIMEOUT_MILLIS,
+                (WatchedEvent event) -> {
+                    synchronized (workerThreadWaitNotifyLock) {
+                        workerThreadWaitNotifyLock.notify();
+                    }
+                });
+        synchronized (workerThreadWaitNotifyLock) {
+            workerThreadWaitNotifyLock.wait(ZOOKEEPER_CONNECTION_TIMEOUT_MILLIS);
+        }
+        ZooKeeper.States state = zooKeeper.getState();
+        if (state == ZooKeeper.States.CONNECTED || state == ZooKeeper.States.CONNECTEDREADONLY) {
+            result = true;
+        }
+        zooKeeper.close();
+        return result;
+    }
+
+    /**
+     * Gets a coordination service for a specific namespace.
      *
      * @param rootNode The name of the root node that defines the namespace.
      *
-     * @return The service for the namespace defined by the root node name.
+     * @return The coordination service.
      *
-     * @throws CoordinationServiceException If an instaNce of the coordination
+     * @throws CoordinationServiceException If an instance of the coordination
      *                                      service cannot be created.
      */
-    public static synchronized CoordinationService getInstance(String rootNode) throws CoordinationServiceException {
+    public static synchronized CoordinationService getServiceForNamespace(String rootNode) throws CoordinationServiceException {
+        /*
+         * Connect to ZooKeeper via Curator.
+         */
         if (null == curator) {
             RetryPolicy retryPolicy = new ExponentialBackoffRetry(1000, 3);
-            // When run in Solr, ZooKeeper defaults to Solr port + 1000
             int zooKeeperServerPort = Integer.valueOf(UserPreferences.getIndexingServerPort()) + PORT_OFFSET;
             String connectString = UserPreferences.getIndexingServerHost() + ":" + zooKeeperServerPort;
             curator = CuratorFrameworkFactory.newClient(connectString, SESSION_TIMEOUT_MILLISECONDS, CONNECTION_TIMEOUT_MILLISECONDS, retryPolicy);
@@ -157,7 +117,7 @@ public final class CoordinationService {
             CoordinationService service;
             try {
                 service = new CoordinationService(rootNode);
-            } catch (Exception ex) {
+            } catch (IOException | InterruptedException | KeeperException | CoordinationServiceException ex) {
                 curator = null;
                 throw new CoordinationServiceException("Failed to create coordination service", ex);
             }
@@ -167,15 +127,18 @@ public final class CoordinationService {
     }
 
     /**
-     * Constructs an instance of the centralized coordination service for a
-     * specific namespace.
+     * Constructs an instance of the coordination service for a specific
+     * namespace.
      *
      * @param rootNodeName The name of the root node that defines the namespace.
+     *
+     * @throws Exception (calls Curator methods that throw Exception instead of
+     *                   more specific exceptions)
      */
-    private CoordinationService(String rootNodeName) throws Exception {
+    private CoordinationService(String rootNodeName) throws InterruptedException, IOException, KeeperException, CoordinationServiceException {
 
         if (false == isZooKeeperAccessible()) {
-            throw new Exception("Unable to access ZooKeeper");
+            throw new CoordinationServiceException("Unable to access ZooKeeper");
         }
 
         String rootNode = rootNodeName;
@@ -191,6 +154,8 @@ public final class CoordinationService {
                 if (ex.code() != KeeperException.Code.NODEEXISTS) {
                     throw ex;
                 }
+            } catch (Exception ex) {
+                throw new CoordinationServiceException("Curator experienced an error", ex);
             }
             categoryNodeToPath.put(node.getDisplayName(), nodePath);
         }
@@ -385,35 +350,77 @@ public final class CoordinationService {
     }
 
     /**
-     * Determines if ZooKeeper is accessible with the current settings. Closes
-     * the connection prior to returning.
-     *
-     * @return true if a connection was achieved, false otherwise
+     * Exception type thrown by the coordination service.
      */
-    private static boolean isZooKeeperAccessible() {
-        boolean result = false;
-        Object workerThreadWaitNotifyLock = new Object();
-        int zooKeeperServerPort = Integer.valueOf(UserPreferences.getIndexingServerPort()) + PORT_OFFSET;
-        String connectString = UserPreferences.getIndexingServerHost() + ":" + zooKeeperServerPort;
+    public final static class CoordinationServiceException extends Exception {
 
-        try {
-            ZooKeeper zooKeeper = new ZooKeeper(connectString, ZOOKEEPER_SESSION_TIMEOUT_MILLIS,
-                    (WatchedEvent event) -> {
+        private static final long serialVersionUID = 1L;
 
-                        synchronized (workerThreadWaitNotifyLock) {
-                            workerThreadWaitNotifyLock.notify();
-                        }
-                    });
-            synchronized (workerThreadWaitNotifyLock) {
-                workerThreadWaitNotifyLock.wait(ZOOKEEPER_CONNECTION_TIMEOUT_MILLIS);
-            }
-            ZooKeeper.States state = zooKeeper.getState();
-            if (state == ZooKeeper.States.CONNECTED || state == ZooKeeper.States.CONNECTEDREADONLY) {
-                result = true;
-            }
-            zooKeeper.close();
-        } catch (InterruptedException | IOException ignored) {
+        private CoordinationServiceException(String message) {
+            super(message);
         }
-        return result;
+
+        private CoordinationServiceException(String message, Throwable cause) {
+            super(message, cause);
+        }
+    }
+
+    /**
+     * An opaque encapsulation of a lock for use in distributed synchronization.
+     * Instances are obtained by calling a get lock method and must be passed to
+     * a release lock method.
+     */
+    public static class Lock implements AutoCloseable {
+
+        /**
+         * This implementation uses the Curator read/write lock. see
+         * http://curator.apache.org/curator-recipes/shared-reentrant-read-write-lock.html
+         */
+        private final InterProcessMutex interProcessLock;
+        private final String nodePath;
+
+        private Lock(String nodePath, InterProcessMutex lock) {
+            this.nodePath = nodePath;
+            this.interProcessLock = lock;
+        }
+
+        public String getNodePath() {
+            return nodePath;
+        }
+
+        public void release() throws CoordinationServiceException {
+            try {
+                this.interProcessLock.release();
+            } catch (Exception ex) {
+                throw new CoordinationServiceException(String.format("Failed to release the lock on %s", nodePath), ex);
+            }
+        }
+
+        @Override
+        public void close() throws CoordinationServiceException {
+            release();
+        }
+    }
+
+    /**
+     * Category nodes are the immediate children of the root node of a shared
+     * hierarchical namespace managed by a coordination service.
+     */
+    public enum CategoryNode {
+
+        CASES("cases"),
+        MANIFESTS("manifests"),
+        CONFIG("config"),
+        RESOURCE("resource");
+
+        private final String displayName;
+
+        private CategoryNode(String displayName) {
+            this.displayName = displayName;
+        }
+
+        public String getDisplayName() {
+            return displayName;
+        }
     }
 }
